Formatted all files, fixed linting errors, refactored ProjectCellContextMenu, moved ProjectCell util functions to its own file
