+++ src/components/Planning/PlanningTable/PlanningRow/DeleteGroupDialog/styles.css
@@ -1,3 +1,3 @@
 #delete-group-dialog {
-    @apply w-auto
+  @apply w-auto;
 }
