+++ src/components/Planning/ProjectProgrammedDialog/styles.css
@@ -17,7 +17,7 @@
 }
 
 .project-suggestions-card {
-  @apply mb-3 border-2 border-black bg-white py-6 px-8;
+  @apply mb-3 border-2 border-black bg-white px-8 py-6;
 }
 
 .project-suggestions-card .search-result-title-container {
