+++ src/components/SearchResults/styles.css
@@ -4,7 +4,7 @@
 }
 
 .search-terms {
-  @apply mr-6 mb-6 flex h-auto flex-wrap gap-2;
+  @apply mb-6 mr-6 flex h-auto flex-wrap gap-2;
 }
 
 /* Tag container (can't be given classes or styles) */
@@ -18,7 +18,7 @@
 }
 
 .search-result-card {
-  @apply mb-3 border-2 border-black bg-white py-6 px-8;
+  @apply mb-3 border-2 border-black bg-white px-8 py-6;
 }
 
 .search-result-card .search-result-title-container {
@@ -52,7 +52,7 @@
 
 /* Search result header */
 .search-result-header-container {
-  @apply pt-16 pl-16 pr-24 pb-6;
+  @apply pb-6 pl-16 pr-24 pt-16;
 }
 
 .search-result-header-container .search-result-page-title {
@@ -95,7 +95,7 @@
 }
 
 .limit-dropdown > div > button > svg {
-  @apply absolute top-0 right-0;
+  @apply absolute right-0 top-0;
 }
 
 .limit-dropdown > div > button > span {
