+++ src/interfaces/common.ts
@@ -75,7 +75,7 @@ export interface ICellMenuDetails {
   cellType: CellType;
   onRemoveCell: () => void;
   onEditCell: () => void;
-  onUpdateCellPhase: (phase:string)=>void;
+  onUpdateCellPhase: (phase: string) => void;
 }
 
 export interface IPhaseMenuDetails {
