+++ src/interfaces/eventInterfaces.ts
@@ -7,7 +7,7 @@ export enum ContextMenuType {
   EDIT_PROJECT_CELL,
   EDIT_PROJECT_PHASE,
   NEW_ITEM,
-  EDIT_GROUP_ROW
+  EDIT_GROUP_ROW,
 }
 
 export interface ICellMenuDetails {
@@ -17,7 +17,7 @@ export interface ICellMenuDetails {
   onRemoveCell: () => void;
   onEditCell: () => void;
   onUpdateCellType: (phase: string) => void;
-  canTypeUpdate:  boolean;
+  canTypeUpdate: boolean;
 }
 
 export interface IPhaseMenuDetails {
@@ -33,15 +33,15 @@ export interface INewItemMenuDetails {
 
 export interface GroupRowMenuDetails {
   groupName: string;
-  onShowGroupDeleteDialog: ()=>void;
-  onShowGroupEditDialog: ()=>void;
+  onShowGroupDeleteDialog: () => void;
+  onShowGroupEditDialog: () => void;
 }
 export interface IContextMenuData {
   menuType: ContextMenuType;
   cellMenuProps?: ICellMenuDetails;
   phaseMenuProps?: IPhaseMenuDetails;
   newItemsMenuProps?: INewItemMenuDetails;
-  groupRowMenuProps?: GroupRowMenuDetails
+  groupRowMenuProps?: GroupRowMenuDetails;
 }
 
 export interface IFinanceEventData {
